XBeeSerialPort: do not open serial port on constructor
The Serial class from the pyserial automatically opens the serial port when
an object is instanciated. This is unconvinient to use, and the current
codebase works around it in the XBeeDevice class.

Move this workaround to the XBeeSerialPort subclass, so that the XBeeDevice
does not need to know about this implementation details of the
communication interface, so that it is easier to generalize.

While on it, remove the unused attribute 'XBeeSerialPort._isOpen'.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>