frame id: fix how the next frame id is calculated
Before this change the frame id was calculated per sent frame and per XBee.
This commit calculates the frames id only per sent frame.

With this change, frames id are independent of the XBee:
   * 1st frame sent: frame id = 1
   * 2nd frame sent: frame id = 2
   ...
   * 255th frame sent: frame id = 255

Before this change, frames id were:
   * In XBee A:
       * For its 1st frame: frame id = 1
       * For its 2nd frame: frame id = 2
       ...
       * For its 255th frame: frame id = 255
   * In XBee B:
       * For its 1st frame: frame id = 1
       * For its 2nd frame: frame id = 2
       ...
       * For its 255th frame: frame id = 255

This is a problem when checking for the corresponding frame reply, because
the source address of a received frame reply was not taken into account and
compared with the destination address in the sent frame.
This may cause problems because a reply frame from B can be associated to the
sent frame for A just because their frame id is the same.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>