devices: return the local XBee in methods 'get_device_by_XXXX()' if
The local XBee device is not included in a 'XBeeNetwork' (or subclasses).
When getting an XBee in the network by its 64-bit address, 16-bit address, or
node identifier, 'None' is returned although the local XBee matches with the
search requirements.

This commit checks the search parameters also for the local XBee.
The local XBee has the higher priority in the search.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>