address: ensure 64-bit and 16-bit addresses are 'UNKNOWN' and not 'None'
If a remote XBee is instanciate with a 64-bit or 16-bit address to 'None', this
commit ensures they are initialized to 'UNKNOWN'.

This is required because the 64-bit address is the identifier of a node (it is
unique per XBee, the rest are variable or may not exist) and it is used for
comparisons between XBee objects.

Another consecuence is a not usable object to, for example, get any parameter
(it throws an exception when the 'RemoteATCommandPacket' is being created)

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>