device: create the proper XBee object (base on protocol) when receiving data
When receiving data from a node not in the network cache a 'RemoteXBeeDevice'
is created without the protocol initialized. This may cause problems when
trying to execute actions that depends on the XBee protocol (such as getting the
route or neighbor tables)

This commit ensures the proper remote XBee object is created
('RemoteRaw802Device', 'RemoteDigiMeshDevice', 'RemoteDigiPointDevice',
'RemoteZigBeeDevice') depending on the local XBee protocol.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>