devices: wait until reader thread fully starts
On slow or highly loaded systems, the Reader thread may not be ready
by the time the _send_packet method is called, resulting in a XBeeException
("Packetlistener is not running").

In order to prevent this, block in the device open() method until the reader
thread is fully started.

This is reproducible in the IX14 XBee prototypes when they are under heavy
load (for example soon after boot).

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>