at command: use the status lowest nibble for local/remote at command responses
This commit uses the least significant nibble of status for ATCommResponsePacket
(0x88) and RemoteATCommandResponsePacket (0x97). The most significant nibble is
ignored.

For example if receiving a 0x40 as status, the value is OK, as it is an 0x00.

The actual hexadecimal value of the status is saved for the frame as the
attribute 'real_state' and used to recreate it. The attribute 'status' returns
the `ATCommandStatus' associated to the low nibble.

https://jira.digi.com/browse/XCTUNG-376

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>