discovery full: define 'NT' limits and allow using already configured value
This commit defines the minumum and maximum value of 'NT' parameter for based
on the local XBee protocol (802.15.4, Zigbee or DigiMesh).

It also allows to use the already configured 'NT' value in the local XBee if
none is provided with 'set_deep_discovery_timeouts()' (for 'node_timeout'
parameter), instead of configuring a default value.
('NT' is used for the discovery process of DigiMesh, 802.15.4, or none-protocol
instances, not for Zigbee)

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>