devices: digimesh: set 16bit address as unknown
On DigiMesh, the 16 bit address is set to 0xFFFE on the XBee frames, which is
the value returned by get_16bit_addr() method on a RemoteXBee.

On local XBee devices, however, 'None' was being returned. This commits
makes it so that 0xFFFE is also returned when using local XBee devices,
fixing this inconsistency.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>