profile: change the way profile is initialized to avoid un-compressing it
The profile parameters and settings are now read without un-compressing the
file to avoid extra processing time and required space when the only goal
is to retrieve profile information.
When requesting profile paths to read files, the profile file is automatically
un-compressed to a temporary folder that is later removed when the object is
disposed like it was done before.

Signed-off-by: David Escalona <david.escalona@digi.com>