firmware ota: use requested file offset and size instead of fixed chunks
When the OTA client (remote XBee) sends a 'next image request' it includes the
offset and a maximum size. Use this received data to read the proper amount
of bytes from the OTA file starting from the received offset and send to the
client (remote).

Before this commit, when the remote XBee asks for a specific offset, the code
assumed that the partition of the firmware in packages was always at fixed
boundaries. For example, if the size received in the 'next image request' was
64 bytes, the ota file was split in chunks of 64 bytes, the received offset
is identified with a chunk index and this complete chunk was sent.

The problem with this implementation happens when the remote asks for an
offset that does not corresponds with the beginning of a chunk, but it is
in the middle of the last chunk sent. Under this situation, the old code would
sent the same chunk previously sent (not the block starting at the specified
offset), so the firmware update process would fail. For example, the first
requested offset is 0 and the size is 64 bytes, then the remote asks for other
64 bytes starting at offset 53. The old code would sent the first 0-63 block
again, instead of the 53-116 block that is being requested.

https://jira.digi.com/browse/XBPL-344

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>