firmware ota: add mechanism to calculate the proper block size
This commit adds a mechanism to calculate the block size of the 'image block
response' packet based on the maximum size received on the 'image block
request'.

The block size to send starts with the maximum received in the request. If a
'payload too large' transmit status is received, the new try decrements the
size 1 byte. This is done until a success transmit status happens.

https://jira.digi.com/browse/XBPL-346

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>