local: return hardcoded protocol for local typed XBee instances when they've never been open
The protocol of any local XBee object (typed or not) is initialized when it is
opened. Before, its value is 'None'.

For typed local classes (ZigBeeDevice, DigiMeshDevice, etc.) we can return a
hardcoded value if the object is not open yet, and then return the initialized
value calculated from the read hardware and firmware version.

This avoids problems when the knowledge of the protocol is required before
opening the device.

https://jira.digi.com/browse/XNM-61

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>