discovery full: zigbee: if error 'address not found', get the node 16-bit addr
While discovering a Zigbee network the error 'address not found' may happen if
the cached 16-bit address is not the right one.

This commit solves this by trying read the new value of the 16-bit address and,
if it is different from the previous value, adding the node again to the FIFO to
look for its neighbors.

https://jira.digi.com/browse/XNM-64

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>