ota firmware: XBee 3 ZB/DM transmit status errors on 'end response' only for certain fw versions
For Zigbee, the 'Address not found' error in the received trasmit status after
sending an 'Upgrade end response' frame only occurs for firmware versions 1009
or lower.

For DigiMesh, the 'No ack' error in the received trasmit status after sending
an 'Upgrade end response' frame only occurs for firmware versions 3004 or lower.

The workaround of considering these errors as valid is limited for firmware
versions 1009 or lower in Zigbee and 3004 or lower in DigiMesh.

https://jira.digi.com/browse/XBPL-345
https://jira.digi.com/browse/XBHAWKDM-796

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>