init XBee: fix update the XBee info when sending an AT command
If using 'set_parameter' method of an XBee object, the check was not being
performed.

This commit moves the check to the AbstractXBeeDevice class so it is always
performed, and the XBee info parameters NI or MY update its value for the
corresponding node.

Related to commit 81716a91dce0d944b65219d9c204dedb0cb5440e

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>