filesystem: profile: do not use ota filesystem image for remote updates
Use the same resources as the local filesystem to update a remote XBee file
system instead of an OTA image for XBee 3 firmware versions greater than x0B.

Currently this means the node filesystem (local or remote) is formatted and then
the embedded file system uploaded. In the future we may allow users to configure
in the profile what they want to keep, overwrite, or upload instead of
formatting the whole filesystem.

https://jira.digi.com/browse/XBPL-363

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>