digimesh: find neighbors: fix minor issues
If no RSSI is included in the response, set it to -9999.
If no neighbor callback is defined do not try to invoke it.

While on it add a comment to explain how "APPEND_DD" and "APPEND_RSSI"
discovery options work: configured discovery options "APPEND_DD" and
"APPEND_RSSI" in the local XBee affects also to the FN requests of the remotes.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>