requirements: move 'srp' package to 'extras_require'
'srp' is an optional feature only required to work with bluetooth, specifically
to change an XBee bluetooth password with method 'update_bluetooth_password()'.

Moving the requirement to 'extras_require' and outside 'install_requires' allows
to install applications without SRP installed, since it is optional and will
not be automatically installed unless another package dependends on it by
including the desired "extras".

https://setuptools.readthedocs.io/en/latest/setuptools.html

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>