profile: do not uncompress a profile to check if it is valid for a serial port
When recovering an XBee using a profile (passing a serial port instead of an
XBeeDevice object to the 'apply_xbee_profile()' method), it must include the
local firmware binary files to be programmed.

This commit avoids to uncompress the whole profile just to check if the profile
is valid for this purpose.

Related to commit 1d57c26ae82bff128e2fca7d77a1f895fe691680

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>