profile: increase the retries for settings update on remote nodes
If network settings are updated on remote nodes to the same value, when applying
the remote node desconects from the network and connects again, that takes
between 10 and 17 seconds (experimental with Zigbee and only one hop).

This commits increases the retries to wait for the 'AC' response and to read
the node information and also adds a time to wait for the node to reconnect.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>