firmware: ota: fix XBee 3 ota image size calculation
Depending on the firmware in the remote XBee being updated the size and the
image to transfer is different:
  * For new firmware versions, x0A and higher, the entire OTA file is sent to
    the remote. This means the size to send in the 'Query Next Image response'
    of the ZCL firmware update protocol is the OTA file size.

  * For older firmware versions (below x0A), only the GBL embedded in the OTA
    file is sent to the remote. This means the size to send in the 'Query Next
    Image response' of the ZCL firmware update protocol is the 'length' field
    of the first sub-element tag (located always at 'Header length' offset)

While on it, added 0x25 transmit status (Route not found) as an error code to
omit when updating from a DigiMesh 3004 or less version when sending an
'Upgrade end response'. Also added a timeout between sent retries of 'Query
next image response' frame, since the same 0x25 (Route not found) is received.

https://jira.digi.com/browse/XCTUNG-2054

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>