route: zigbee: force getting route to node by setting again the AR parameter
When AR != FF in the local XBee (concentrator), remote nodes sends Route Record
Indicator frames to the concentrator:

   * If AR = 0, they send a Router Record Indicator just once (when they have to
     send anything to the concentrator).
   * If AR > 0 and AR < FF, they send a Route Record Indicator once during the
     configured interval, AR*10 seconds (when they have to send anything to the
     concentrator)

If the AR value is refreshed with the same value, the interval is reset and
remotes send again a Route Record Indicator. That is, by refreshing AR value in
the concentrator a new Route Record Indicator from any remote can be forced by
requesting/sending data to them.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>