models/hw: add new enum listing the legacy hardware versions based on letters
Initial XBee firmware XML description files used letters instead of hexadecimal
values to represent a specific hardware version. This enum is a mapping of those
letters with their corresponding hardware version. This way, if a letter is
returned instead of an hexadecimal number while reading the hardware version from
an XML file, this enum can be used as fallback to retrieve the correct hardware
version value.

Signed-off-by: David Escalona <david.escalona@digi.com>