devices: remote at: check received 16-bit address is known
This commit checks the source 16-bit address of the received 'Remote AT command
response' frames to ensure it is a known address (not 'FFFE') before comparing
with the one in the sent 'Remote AT command request' as destination.

This could happen if the node to request is not in the network anymore.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>