get: set: execute: add a new optional arg to set/get/execute parameter methods
New optional argument 'apply' for methods:
   * 'get_parameter()'
   * 'set_parameter()'
   * 'execute_parameter()'

This way internal executions can just use it and do not need to read the
general apply flag ('is_apply_changes_enabled()') to update it
('enable_apply_changes()') and then restore the original value.

Default value for 'apply' is 'None' to keep the behaviour the method had in
previous versions.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>