discover: keep node 'packet_received_from' callbacks if not completely removed
When a deep discovery process finishes, it purges nodes and connections in the
network that were not discovered. This process implies removing the objects if
deep discovery option 'del_not_discovered_nodes_in_last_scan' is 'True'.
If this option is 'False' nodes are just labelled as unreachable but the objects
stay in the network list.

In this case, all registered 'packet_received_from' callbacks of a node must
not be removed. This should only occur when completely deleting not discovered
nodes ('del_not_discovered_nodes_in_last_scan=True')

https://jira.digi.com/browse/XBPL-355

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>