update: firmware: profile: set end devices to sleep the minimum while updating
This commit configures 'SP' (sleep period) and 'SN' (number of sleep periods) to
their minimum possible value (0x20 for 'SP' and 1 for 'SN') so an end device is
reachable.

After a profile update, the right values are restored.

For a firmware update, initial configuration is not restored since the values
are not written to flash and a reboot of the node is performed once the process
finishes.

https://jira.digi.com/browse/DAL-4158

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>