recovery: enter in command mode when recovering the local XBee
If a lot of data is being output in the serial port, the XBee connection cannot
be recovered.

This commit forces the command mode before executing the commands to configure
the XBee connection.
It also swaps the reader stop functionality to first mark the thread to stop
and then try to quit the reading process of the communication interface. If not,
the open process stucks waiting for valid frames in the communication interface.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>