ota: s2c: fix order updater candidates
The node to be updated can be 'node_a' or 'node_b' of a connection, that only
depends on which of the nodes was discovered first (this one is 'node_a').

While on it simplify the code to use just one method to get updater candidates
with an argument to use already discovered connections or perform a neighbor
discover.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>