refresh: digimesh: reset 16-bit address for protocols that does not support it
When a Zigbee local node is updated to be a DigiMesh node, its 16-bit address
is not reset although its information is forced to be re-initalized. This
commit forces the 16-bit address to be 0xFFFE (unknown) for any protocol not
supporting 16-bit addressing.

Related to cc685b12eae508714703ac804d8584b9035ca0bd

https://jira.digi.com/browse/XNM-87

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>