encoding: use utf-8 to encode/decode and ignore malformed data
When encode or decode using utf-8 encoding avoid any 'UnicodeError' just
ignore the malformed data.

Malformed data may be received for example when working in DigiMesh encrypted.
If 2 nodes used different encryption keys (KY) data is received corrupted since
the receiver is unencrypting with the wrong key. This data may not be utf-8 in
fields that are expected to be, for example the AT command of a Remote AT
Command frame.

https://jira.digi.com/browse/XBHAWKDM-901

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>