update: profile: do not pass the bootloader file when updating a remote profile
The bootloader file attribute in the 'XBeeProfile' attribute is the '*.gbl'
file to be applied locally. For remote updates the '*.otb' file is required.

If not bootloader file is provided, the library looks for a '*.otb' file next
to the firmware '*.xml' file (the same it does for the corresponding '*.ota' if
not provided)

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>