firmware update: ota: do not wait for transmit status if first block request arrives
If, after sending the Query Image Next Image Response, the first Image Block
Request arrives before its Trasmit Status, ignore this missing frame and
continue with the process sending the corresponding Image Block Response.

This mechanism was already implemented for the Image Block Responses in
commit 2016bfb0bf27ad5469e5b2c9f231bd9719f88a2c.

While on it fix some typos in comments.

https://onedigi.atlassian.net/browse/XBPL-377

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>