profile: fix protocol change after applying a profile in SX devices
In SX devices, protocol changes are not linked to the firmware file applied.
SX devices can change the protocol by updating the value of the BR parameter.
Checking whether the protocol has changed after applying the settings is
mandatory to avoid future failures with XBee device class instances. This
commit throws an exception if, after updating the profile settings of an SX
device, the protocol of the device has changed, allowing upper layers to
re-instantiate the correct XBee device class.

Signed-off-by: David Escalona <david.escalona@digi.com>