devices: add BR property to AbstractXBeeDevice
SX devices require the BR value to correctly initialize the device protocol.
This value can be read directly from the device when the communication
interface is available, but in some scenarios the communication interface is busy
and the BR value cannot be read, resulting in an error initializing the device.
This commit adds the BR value as a new property that is cached in the device once
it is read, allowing upper layers to retrieve it in order to initialize the device
protocol when the communication interface is busy.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>
Signed-off-by: David Escalona <david.escalona@digi.com>