explicit data: add explicit data to 802.15.4
XBee 3 802.15.4 modules allows to send and receive explicit data (other module
hardware does not include this feature, i. e., S2C)

This commits includes methods to read and send explicit data to the 'XBeeDevice'
class so they are inherited by 'Raw802Device'. It removes the equivalent methods
from 'ZigBeeDevice', 'DigiMeshDevice', and 'DigiPointDevice' classes since now
they are defined in the super class. It also overrides the methods in the
'IPDevice' class so they throw an exception if used as they are not supported
for Cellular nor Wi-Fi XBee modules.

The related documentation is also updated.

https://onedigi.atlassian.net/browse/DAL-6334

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>