firmware: fix AttributeError when update_local_firmware is called with a string
This finally block assumed that target was always an XBeeDevice,
while all of the code above it did the right thing and properly
handled the case of it being a str.

The "Bootloader not supported" block also assumed target was
always an XBeeDevice, so that code has been fixed as well.