srp: remove 'pysrp' dependency
'pysrp' library includes 2 modules to implement SRP:

   * _ctsrp: Using OS libraries to implement SRP.
   * _pysrp: Pure Python implementation of SRP.

The first option is always '_ctsrp'. If there is any problem loading the
required libraries or looking for required methods inside, '_pysrp' is loaded.
This last module, '_pysrp' has a bug in the generation of the verifier for
certain combination of username and password (for example, user=apiservice,
password=1257c) See https://github.com/cocagne/pysrp/issues/55.

This commit removes the dependency to 'pysrp' to avoid this bug and includes
the required implementation to calculate a salt and the corresponding verifier
to configure XBee Bluetooth password.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>